+++ FindPanel/OgreTextFindComponents/NSOutlineView/OgreOutlineItemAdapter.m
@@ -99,17 +99,17 @@
 - (BOOL)isHighlightable { return NO; }
 
 /* Getting structural detail */
-- (unsigned)numberOfChildrenInSelection:(BOOL)inSelection
+- (NSUInteger)numberOfChildrenInSelection:(BOOL)inSelection
 {
 #ifdef DEBUG_OGRE_FIND_PANEL
 	NSLog(@"  -numberOfChildrenInSelection: of %@", [self className]);
 #endif
-    unsigned    count = [_outlineColumn ogreNumberOfChildrenOfItem:_item];
+    NSUInteger  count = [_outlineColumn ogreNumberOfChildrenOfItem:_item];
     
     return 1 /* self cell */ + count;
 }
 
-- (id)childAtIndex:(unsigned)index inSelection:(BOOL)inSelection
+- (id)childAtIndex:(NSUInteger)index inSelection:(BOOL)inSelection
 {
 #ifdef DEBUG_OGRE_FIND_PANEL
 	NSLog(@"  -childAtIndex:%d of %@", index, [self className]);
@@ -170,7 +170,7 @@
 #ifdef DEBUG_OGRE_FIND_PANEL
 	NSLog(@"  -selectedIndexes of %@", [self className]);
 #endif
-    unsigned count = [self numberOfChildrenInSelection:YES];
+    NSUInteger count = [self numberOfChildrenInSelection:YES];
     
     return [NSIndexSet indexSetWithIndexesInRange:NSMakeRange(0, count)];
 }
@@ -204,12 +204,12 @@
     return _outlineColumn;
 }
 
-- (void)setLevel:(int)level
+- (void)setLevel:(NSInteger)level
 {
     _level = level;
 }
 
-- (int)level
+- (NSInteger)level
 {
     return _level;
 }
