+++ FindPanel/OgreTextFindComponents/NSTableView/OgreTableCellAdapter.h
@@ -20,9 +20,9 @@
 @interface OgreTableCellAdapter : OgreTextFindLeaf
 {
     OgreTableColumn *_tableColumn;
-    int             _rowIndex;
+    NSInteger       _rowIndex;
 }
 
-- (id)initWithTableColumn:(OgreTableColumn*)tableColumn row:(int)rowIndex;
+- (id)initWithTableColumn:(OgreTableColumn*)tableColumn row:(NSInteger)rowIndex;
 
 @end
