+++ FindPanel/OgreTextFindComponents/NSTextView/OgreTextViewUndoer.m
@@ -15,7 +15,7 @@
 
 
 @implementation OgreTextViewUndoer
-- (id)initWithCapacity:(unsigned)aCapacity
+- (id)initWithCapacity:(NSUInteger)aCapacity
 {
     self = [super init];
     if (self != nil) {
@@ -54,8 +54,8 @@
 	NSTextStorage       *textStorage = [aTarget textStorage];
     NSRange             aRange, newRange;
     NSAttributedString  *aString;
-    unsigned            i;
-    OgreTextViewUndoer    *redoArray = [[OgreTextViewUndoer alloc] initWithCapacity:_count];
+    NSUInteger          i;
+    OgreTextViewUndoer  *redoArray = [[OgreTextViewUndoer alloc] initWithCapacity:_count];
     
     [textStorage beginEditing];
