+++ FindPanel/OgreTextFindComponents/OgreTextFindBranch.m
@@ -99,7 +99,7 @@
 - (BOOL)isLeaf { return NO; }
 - (BOOL)isBranch { return YES; }
 
-- (unsigned)numberOfChildrenInSelection:(BOOL)inSelection
+- (NSUInteger)numberOfChildrenInSelection:(BOOL)inSelection
 {
 #ifdef DEBUG_OGRE_FIND_PANEL
 	NSLog(@"  -numberOfChildrenInSelection: of %@ (BUG!!!)", [self className]);
@@ -107,9 +107,9 @@
     return 0; 
 }
 
-- (unsigned)numberOfDescendantsInSelection:(BOOL)inSelection
-{ 
-    unsigned            numberOfDescendants = 0;
+- (NSUInteger)numberOfDescendantsInSelection:(BOOL)inSelection
+{
+    NSUInteger          numberOfDescendants = 0;
     NSEnumerator        *enumerator = (NSEnumerator*)[self componentEnumeratorInSelection:inSelection];
     OgreTextFindLeaf    *aChild;
     
@@ -124,7 +124,7 @@
     return numberOfDescendants;
 }
 
-- (id)childAtIndex:(unsigned)index inSelection:(BOOL)inSelection
+- (id)childAtIndex:(NSUInteger)index inSelection:(BOOL)inSelection
 {
 #ifdef DEBUG_OGRE_FIND_PANEL
 	NSLog(@"  -childAtIndex: of %@ (BUG!!!)", [self className]);
@@ -178,12 +178,12 @@
     _isParentRetained = NO;
 }
 
-- (int)index
+- (NSInteger)index
 {
     return _index;
 }
 
-- (void)setIndex:(int)index
+- (void)setIndex:(NSInteger)index
 {
     _index = index;
 }
